10 more lines of comments / 2 less lines of code
I was trying to understand this, but i have to go to bed.
so this is me trying to understand stuff

no functional change. well i did use an f-string (needs python 3.6). and less duplicate code
this is a way of me sharing code. you don't have to "merge" or anything